fix NullException

+++ src/Reenbit.ChuckNorris.Services/JokeService.cs
@@ -155,8 +155,12 @@ namespace Reenbit.ChuckNorris.Services
             using (IUnitOfWork uow = unitOfWorkFactory.CreateUnitOfWork())
             {
                 var jokeRepository = uow.GetRepository<IJokeRepository>();
-                var favoriteJokeForDelete = await jokeRepository.FindUserFavoriteAsync(uf => uf.UserId == Int32.Parse(userId) && uf.JokeId == favoriteJokeId);
-                jokeRepository.RemoveUserFavorite(favoriteJokeForDelete);
+                var favoriteJokeForDelete = (await jokeRepository.FindUserFavoriteAsync(uf => uf.UserId == Int32.Parse(userId) && uf.JokeId == favoriteJokeId));
+                if (favoriteJokeForDelete != null)
+                {
+                    jokeRepository.RemoveUserFavorite(favoriteJokeForDelete);
+                }
+
                 var number = await (uow.SaveChangesAsync());
                 return number > 0;
             }
